requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
pyarrow>=14.0.0
tensorflow>=2.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
